# swagger/swagger.yml
swagger: "2.0"
info:
  description: "IPFS api Rest"
  title: "ipfs-api"
  version: "v1"
host: "localhost:4040"
schemes:
  - "http"
paths:
  /file:
    post:
      tags:
      - "file"
      summary: "Post ipfs file"
      description: "Addd the content to ipfs"
      operationId: "postFile"
      consumes:
      - application/json
      parameters:
      - name: "content"
        in: "body"
        description: "File content"
        schema:
          type: object
          required: 
          - content
          properties:
            content:
              type: string
      produces:
      - "application/json"
      responses:
        201:
          description: "Successful operation"
          schema:
            type: object
            properties:
              error:
                type: boolean
              code:
                type: integer
              cid:
                type: string
              message:
                type: string
        '500':
          description: "Internal error"
  /file/{cid}:
    get:
      tags:
      - "file"
      summary: "Get ipfs file"
      description: "Returns the content of the file"
      operationId: "getFile"
      produces:
      - "application/json"
      parameters:
      - name: "cid"
        in: "path"
        required: true
        type: string
      responses:
        '200':
          description: "Successful operation"
          schema:
            type: "object"
            properties:
              error: 
                type: boolean
              code: 
                type: integer
              message:
                type: string
              content:
                type: string
        '400':
          description: "Invalid cid"
          schema:
            type: object
            properties:
              error: 
               type: boolean
              code:
               type: integer
              message:
               type: string
        '500':
          description: "Internal error"
  /file/ipns/{cid}:
    get:
      tags:
      - "file"
      summary: "Get ipfs file"
      description: "Returns the content of the file"
      operationId: "getIpnsFile"
      produces:
      - "application/json"
      parameters:
      - name: "cid"
        in: "path"
        required: true
        type: string
      responses:
        '200':
          description: "Successful operation"
          schema:
            type: object
            properties:
              error: 
                type: boolean
              code: 
                type: integer
              message:
                type: string
              content:
                type: string
        '400':
          description: "Invalid cid"
          schema:
            type: object
            properties:
              error: 
               type: boolean
              code:
               type: integer
              message:
               type: string
        '500':
          description: "Internal error"
  /file/ipns:
    post:
      tags:
      - "file"
      summary: "Post ipfs file"
      description: "Addd the content to ipfs"
      operationId: "postIpnsFile"
      consumes:
      - application/json
      parameters:
      - name: "content"
        in: "body"
        description: "File content"
        schema:
          type: object
          required: 
          - cid
          properties:
            cid:
              type: string
            key:
              type: string
      produces:
      - "application/json"
      responses:
        201:
          description: "Successful operation"
          schema:
            type: object
            properties:
              error:
                type: boolean
              code:
                type: integer
              cid:
                type: string
              message:
                type: string
        '500':
          description: "Internal error"
  /peer:
    get:
      tags:
      - "peer"
      summary: "Get connected peers"
      description: "Returns peers connected to our node"
      operationId: "getPeers"
      produces:
      - "application/json"
      responses:
        '200':
          description: "Successful operation"
          schema:
            type: "object"
            properties:
              error:
                type: boolean
              code:
                type: integer
              peers:
                type: array
                items:
                  type: object
                  properties:
                    addr:
                      type: string
                    peer:
                      type: string
        '500':
          description: "Internal error"
    post:
      tags:
      - "peer"
      summary: "Establish connection"
      description: "Make connection with other peer"
      operationId: "postConnection"
      consumes:
      - application/json
      parameters:
      - name: "addr"
        in: "body"
        description: "Peer address"
        schema:
          type: object
          required: 
          - content
          properties:
            addr:
              type: string
            id:
              type: string
      produces:
      - "application/json"
      
      responses:
        '201':
          description: "Successful operation"
          schema:
            type: object
            properties:
              error:
                type: boolean
              code:
                type: integer
              cid:
                type: string
              message:
                type: string
        '400':
          description: "Incorrect param"
          schema:
            type: object
            properties:
              error:
                type: boolean
              code:
                type: integer
              message:
                type: string
        '500':
          description: "Internal error"
  /peer/{id}:
    get:
      tags:
      - "peer"
      summary: "Get peer by id"
      description: "Returns information about one node"
      operationId: "getPeerInfo"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: string
      responses:
        '200':
          description: "Successful operation"
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: integer
              peerInfo:
                type: object
                properties:
                  id: 
                    type: string
                  publicKey: 
                    type: integer
                  addresses:
                    type: array
                    items:
                      type: string
                  agentVersion:
                    type: string
                  protocolVersion:
                    type: string
                  protocols:
                    type: array
                    items:
                      type: string
        '400':
          description: "Invalid id"
          schema:
            type: object
            properties:
              error: 
               type: boolean
              code:
                type: integer
              message:
                type: string
        '500':
          description: "Internal error"
    delete:
      tags:
      - "peer"
      summary: "Remove peer connection"
      description: "Removes peer connection"
      operationId: "deleteConnection"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: string
      responses:
        '200':
          description: "Successful operation"
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: integer
              message:
                type: string
        '400':
          description: "Invalid id"
          schema:
            type: object
            properties:
              error: 
               type: boolean
              code:
                type: integer
              message:
                type: string
        '500':
          description: "Internal error"
  /pin:
    post:
      tags:
      - "pin"
      summary: "Creates a local pin"
      description: "Pins certain file"
      operationId: "postPin"
      consumes:
      - application/json
      parameters:
      - name: "cid"
        in: "body"
        description: "File identifier"
        schema:
          type: object
          required: 
          - cid
          properties:
            cid:
              type: string
      produces:
      - "application/json"
      responses:
        '201':
          description: "Successful operation"
          schema:
            type: object
            properties:
              error:
                type: boolean
              code:
                type: integer
              message:
                type: string
        '400':
          description: "Invalid cid"
          schema:
            type: object
            properties:
              error:
                type: boolean
              code:
                type: integer
              message:
                type: string
        '500':
          description: "Internal error"
  /pin/{cid}:
    get:
      tags:
      - "pin"
      summary: "Get pin information"
      description: "Returns the information related to the file pinned"
      operationId: "getPin"
      produces:
      - "application/json"
      parameters:
      - name: "cid"
        in: "path"
        description: "File identifier"
        required: true
        type: string
      responses:
        '200':
          description: "Successful operation"
          schema:
            type: "object"
            properties:
              error: 
                type: boolean
              code: 
                type: integer
              pinInfo:
                type: object
                properties:
                  cid: 
                    type: string
                  type: 
                    type: string
        '400':
          description: "Invalid cid"
          schema:
            type: object
            properties:
              error: 
               type: boolean
              code:
                type: integer
              message:
                type: string            
        '404':
          description: "Pin not found"
          schema:
            type: object
            properties:
              error: 
               type: boolean
              code:
                type: integer
              message:
                type: string
        '500':
          description: "Internal error"
    delete:
      tags:
      - "pin"
      summary: "Remove pin"
      description: "Removes the pinning from certain file"
      operationId: "deletePin"
      produces:
      - "application/json"
      parameters:
      - name: "cid"
        in: "path"
        description: "File identifier"
        required: true
        type: string
      responses:
        '200':
          description: "Successful operation"
          schema:
            type: "object"
            properties:
              error: 
                type: boolean
              code: 
                type: integer
              message:
                type: string
        '400':
          description: "Invalid cid"
          schema:
            type: object
            properties:
              error: 
               type: boolean
              code:
                type: integer
              message: 
                type: string
        '404':
          description: "Pin not found"
          schema:
            type: object
            properties:
              error: 
               type: boolean
              code:
                type: integer
              message: 
                type: string
        '500':
          description: "The operation could not be completed."
  /pubsub:
    post:
      tags:
      - "pubsub"
      summary: "Subscribe to certain topic"
      description: "subscription to topic content"
      operationId: "subTopic"
      consumes:
      - application/json
      parameters:
      - name: "content"
        in: "body"
        description: "Topic"
        schema:
          type: object
          required: 
          - topic
          properties:
            topic:
              type: string
      produces:
      - "application/json"
      responses:
        201:
          description: "Successful operation"
          schema:
            type: object
            properties:
              error:
                type: boolean
              code:
                type: integer
              message:
                type: string
        '500':
          description: "the operation could not be completed"
  /pubsub/{topic}:
    post:
      tags:
      - "pubsub"
      summary: "Publish to certain topic"
      description: "Publish data to certain topic"
      operationId: "pubTopic"
      consumes:
      - application/json
      produces:
      - "application/json"
      parameters:
      - name: "topic"
        in: "path"
        description: "Topic"
        required: true
        type: string
      - name: "content"
        in: "body"
        description: "Content to publish"
        schema:
          type: object
          properties:
            content:
              type: string
      responses:
        '201':
          description: "Successful operation"
          schema:
            type: "object"
            properties:
              error: 
                type: boolean
              code: 
                type: integer
              message:
                type: string
        '500':
          description: "The operation could not be completed"
...